Conflicts:
	bower.json

+++ app/scripts/services/code-mirror.js
@@ -34,6 +34,11 @@ angular.module('codeMirror', ['raml', 'ramlEditorApp', 'codeFolding'])
       var cursor = cm.getCursor(), line = cm.getLine(cursor.line),
           indentUnit = cm.getOption('indentUnit'), spaces, result, unitsToIndent;
 
+      if (cm.somethingSelected()) {
+        cm.indentSelection('add');
+        return;
+      }
+
       result = service.removeTabs(line, indentUnit);
       result = result.length ? result : '';
 
@@ -162,7 +167,7 @@ angular.module('codeMirror', ['raml', 'ramlEditorApp', 'codeFolding'])
 
       CodeMirror.registerHelper('hint', 'yaml', ramlHint.autocompleteHelper);
       CodeMirror.registerHelper('fold', 'indent', getFoldRange);
-      
+
       if (!shouldEnableFoldGutter) {
         foldGutterConfig = false;
       } else {
@@ -185,7 +190,8 @@ angular.module('codeMirror', ['raml', 'ramlEditorApp', 'codeFolding'])
         extraKeys: {
           'Ctrl-Space': 'autocomplete',
           'Cmd-s': 'save',
-          'Ctrl-s': 'save'
+          'Ctrl-s': 'save',
+          'Shift-Tab': 'indentLess'
         },
         keyMap: 'tabSpace',
         foldGutter: foldGutterConfig,
